refactor: extract TrackerConnector to a separate file
